PBS:implement xong PBS

+++ kernel/proc.h
@@ -97,8 +97,8 @@ struct proc {
   int stime;   // time spent in SLEEPING state (ticks) //// ref
   uint64 ctime;     // creation time in ticks (for FCFS & tie-breaking)
   uint64 etime;     // nếu m cần time kết thúc, có thể thêm
-  int pbs_rtime;
-  int pbs_stime;
+  uint64 pbs_rtime;   // run time in PBS (ticks)
+  uint64 pbs_stime;   // sleep time in PBS (ticks)
 
 //////////////////////////////////////////////////////////////////////////////////
